requirements.txt
#alembic==1.4.3
#Flask-Cors==3.0.8
flask-marshmallow==0.14.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
marshmallow-sqlalchemy==0.24.1
flask-rest-paginate==1.1.1
pytest==6.2.2
requests==2.25.1
#marshmallow-sqlalchemy==0.23.1
#pipenv==2020.6.2
#pytest==5.4.3
#python-dotenv==0.13.0
#sqlalchemy-mixins==1.2.1
#SQLAlchemy-Utils==0.36.8
#tox==3.15.2
#psycopg2-binary==2.8.5
#bcrypt==3.2.0
